+++ ampd/src/sui/json_rpc.rs
@@ -1,10 +1,9 @@
 use async_trait::async_trait;
-use error_stack::Report;
 use ethers::providers::{JsonRpcClient, ProviderError};
 use sui_json_rpc_types::{SuiTransactionBlockResponse, SuiTransactionBlockResponseOptions};
 use sui_types::digests::TransactionDigest;
 
-use crate::evm::json_rpc::Client;
+use crate::json_rpc::Client;
 
 type Result<T> = error_stack::Result<T, ProviderError>;
 
@@ -25,19 +24,16 @@ where
         &self,
         digests: Vec<TransactionDigest>,
     ) -> Result<Vec<SuiTransactionBlockResponse>> {
-        self.provider
-            .request(
-                "sui_multiGetTransactionBlocks",
-                (
-                    digests
-                        .iter()
-                        .map(|d| d.base58_encode())
-                        .collect::<Vec<_>>(),
-                    SuiTransactionBlockResponseOptions::new().with_events(),
-                ),
-            )
-            .await
-            .map_err(Into::into)
-            .map_err(Report::from)
+        self.request(
+            "sui_multiGetTransactionBlocks",
+            (
+                digests
+                    .iter()
+                    .map(|d| d.base58_encode())
+                    .collect::<Vec<_>>(),
+                SuiTransactionBlockResponseOptions::new().with_events(),
+            ),
+        )
+        .await
     }
 }
